chatbot/requirements.txt
rasa[spacy]==1.3.9
mecab-python3